+++ example/integration_test/plugin_integration_test.dart
@@ -10,7 +10,7 @@
 import 'package:flutter_test/flutter_test.dart';
 import 'package:integration_test/integration_test.dart';
 
-import 'package:floating_panel/floating_panel.dart';
+import 'package:macos_floating_panel/floating_panel.dart';
 
 void main() {
   IntegrationTestWidgetsFlutterBinding.ensureInitialized();